\name{getAllFactors}
\alias{getAllFactors}
\title{Get names of all factors in a transsys frame}
\description{
  Get names of all factors in a transsys frame
}
\usage{
getAllFactors(transsysFrame)
}
\arguments{
  \item{transsysFrame}{the transsys frame}
}
\value{
  The names of all factors, as a character vector.
}
\references{\url{http://www.cmp.uea.ac.uk/~jtk/transsys/}}
\author{Jan T. Kim, \email{jtk@cmp.uea.ac.uk}}
\examples{
\dontrun{
getAllFactors(tp);
}
}
\keyword{misc}
