\name{transsys-package}
\alias{transsys-package}
\alias{transsys}
\docType{package}
\title{transsys Utilities}
\description{
  Utilities for analysing transsys data with R
}
\details{
\tabular{ll}{
Package: \tab transsys\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2006-07-18\cr
License: \tab GPL version 2 or newer\cr
}
}
\keyword{package}
\references{\url{http://www.cmp.uea.ac.uk/~jtk/transsys/}}
\author{Jan T. Kim, \email{jtk@cmp.uea.ac.uk}}
